Move functions parsing 1 Token (primitives) into own file
